+++ CHANGES.md
@@ -1,8 +1,40 @@
-Version 2.0.1
+Version 2.2.0
 -------------
 
+* Add Kafka 0.9.0.1 compatibility. `BPSO-17091`.
+
+Version 2.1.1
+-------------
+
+* Switch to warnings.warn for failure to import native-code Murmur
+  hash `BPSO-13212`.
+
+Version 2.1.0
+-------------
+
+* Fixed bug where Afkak would raise a KeyError when a commit failed
+  `BPSO-11306`
+* Added `request_retry_max_attempts` parameter to Consumer objects
+  which allows the caller to configure the maximum number of attempts
+  the Consumer will make on any request to Kafka. `BPSO-10531`
+  NOTE: It defaults to zero which indicates the Consumer should retry
+  forever. *This is a behavior change*.
+* If a user-initiated Commit operation is attempted while a commit is
+  ongoing (even a Consumer auto-initiated one), the new attempt will
+  fail with a OperationInProgress error which contains a deferred
+  which will fire when the previous Commit operation completes. This
+  deferred *should* be used to retry the Commit since the ongoing
+  operation may not include the latest offset at the time the second
+  operation was initiated.
+* Fixed an error where Commit requests would only be retried once
+  before failing.
+* Reduced the level of some log messages from ERROR to WARNING or
+  lower. `BPSO-11309`
 * Added `update_cluster_hosts` method to allow retargeting Afkak
-  client in case all brokers are restarted on new IPs.
+  client in case all brokers are restarted on new IPs. `BPSO-3521`
+* Fixed bug where Afkak would continue to try to contact brokers at IP
+  addresses they no longer listened on, or brokers which had been
+  removed from the cluster. `BPSO-6790`
 
 Version 2.0.0
 -------------